+++ src/components/layout/CardMain.tsx
@@ -1,6 +1,6 @@
 export default function CardMain({ children }: { children: React.ReactNode }) {
     return (
-        <main className="mx-auto w-full max-w-4xl my-15 border border-border bg-bg rounded-3xl">
+        <main className="mx-auto w-full md:max-w-4xl md:my-15 md:border md:border-border bg-bg md:rounded-3xl">
             {children}
         </main>
     );